AI-Content-Generator/requirements.txt
streamlit==1.28.0
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
python-dotenv==1.0.0